requirements.txt
streamlit
langchain
langchain-community
langchain-groq
faiss-cpu
sentence-transformers
arxiv
pymupdf
python-dotenv